requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
requests==2.32.3
setuptools==79.0.1
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.9.0
openai>=1.0.0